requirements.txt
browser-use
langchain-google-genai
google-genai
mss
Pillow
SpeechRecognition
sounddevice
soundfile
pyperclip
python-dotenv
